requirements.txt
paho-mqtt
pyyaml
requests
